requirements.txt
Django==2.1.3
djangorestframework==3.9.0

asgiref==3.4.1
dj-database-url==0.5.0
django-filter
django-heroku==0.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0